+++ core/common/TimeUtil.h
@@ -79,7 +79,9 @@ void UpdateTimeDelta(time_t serverTime);
 
 uint64_t GetPreciseTimestampFromLogtailTime(LogtailTime logTime, const PreciseTimestampConfig& preciseTimestampConfig);
 
-void SetLogTime(sls_logs::Log* log, time_t second, long nanosecond, bool ignoreNanosecond = false);
+void SetLogTime(sls_logs::Log* log, time_t second);
+
+void SetLogTime(sls_logs::Log* log, time_t second, long nanosecond);
 
 LogtailTime GetCurrentLogtailTime();