+++ core/models/LogEvent.cpp
@@ -73,9 +73,10 @@ Json::Value LogEvent::ToJson() const {
 }
 
 bool LogEvent::FromJson(const Json::Value& root) {
-    SetTimestamp(root["timestamp"].asInt64());
     if (root.isMember("timestampNanosecond")) {
-        SetTimestampNanosecond(root["timestampNanosecond"].asInt64());
+        SetTimestamp(root["timestamp"].asInt64(), root["timestampNanosecond"].asInt64());
+    } else {
+        SetTimestamp(root["timestamp"].asInt64());
     }
     if (root.isMember("contents")) {
         Json::Value contents = root["contents"];