+++ core/models/MetricEvent.cpp
@@ -37,9 +37,6 @@ Json::Value MetricEvent::ToJson() const {
 
 bool MetricEvent::FromJson(const Json::Value& root) {
     SetTimestamp(root["timestamp"].asInt64());
-    if (root.isMember("timestampNanosecond")) {
-        SetTimestampNanosecond(root["timestampNanosecond"].asInt64());
-    }
     return true;
 }