+++ pkg/helper/log_helper.go
@@ -50,7 +50,7 @@ func CreateLog(t time.Time, configTag map[string]string, logTags map[string]stri
 		}
 		slsLog.Contents = append(slsLog.Contents, cont)
 	}
-	protocol.SetLogTime(&slsLog, uint32(t.Unix()), uint32(t.Nanosecond()))
+	protocol.SetLogTime(&slsLog, uint32(t.Unix()))
 	return &slsLog, nil
 }
 
@@ -85,7 +85,7 @@ func CreateLogByArray(t time.Time, configTag map[string]string, logTags map[stri
 		}
 		slsLog.Contents = append(slsLog.Contents, cont)
 	}
-	protocol.SetLogTime(&slsLog, uint32(t.Unix()), uint32(t.Nanosecond()))
+	protocol.SetLogTime(&slsLog, uint32(t.Unix()))
 	return &slsLog, nil
 }
 
@@ -150,7 +150,7 @@ func (hd *HistogramData) ToMetricLogs(name string, timeMs int64, labels MetricLa
 func NewMetricLog(name string, timeMs int64, value string, labels []MetricLabel) *protocol.Log {
 	strTime := strconv.FormatInt(timeMs, 10)
 	metric := &protocol.Log{}
-	protocol.SetLogTime(metric, uint32(timeMs/1000), uint32((timeMs*1e6)%1e9))
+	protocol.SetLogTime(metric, uint32(timeMs/1000))
 	metric.Contents = []*protocol.Log_Content{}
 	metric.Contents = append(metric.Contents, &protocol.Log_Content{Key: "__name__", Value: name})
 	metric.Contents = append(metric.Contents, &protocol.Log_Content{Key: "__time_nano__", Value: strTime})