+++ plugins/input/skywalkingv3/ot_trace.go
@@ -69,10 +69,10 @@ type OtSpanRef struct {
 func (ot *OtSpan) ToLog() (*protocol.Log, error) {
 	log := &protocol.Log{}
 	if ot.End != 0 {
-		protocol.SetLogTime(log, uint32(ot.End/int64(1000000)), uint32((ot.End*1000)%1e9))
+		protocol.SetLogTimeWithNano(log, uint32(ot.End/int64(1000000)), uint32((ot.End*1000)%1e9))
 	} else {
 		nowTime := time.Now()
-		protocol.SetLogTime(log, uint32(nowTime.Unix()), uint32(nowTime.Nanosecond()))
+		protocol.SetLogTimeWithNano(log, uint32(nowTime.Unix()), uint32(nowTime.Nanosecond()))
 	}
 	log.Contents = make([]*protocol.Log_Content, 0)
 	linksJSON, err := json.Marshal(ot.Links)